fix(ci): resolve PHPStan false positives in static analysis bootstrap
Stop forcing CLI_SCRIPT in phpstan-bootstrap so runtime guards in hooks and
admin settings analyse as web context, and cast XMLDB_INDEX_UNIQUE for add_index.

+++ db/upgrade.php
@@ -101,7 +101,7 @@ function xmldb_local_artqtml_upgrade($oldversion) {
         $table->add_field('hitcount', XMLDB_TYPE_INTEGER, '10', null, XMLDB_NOTNULL, null, '0');
         $table->add_key('primary', XMLDB_KEY_PRIMARY, ['id']);
         $table->add_key('userid', XMLDB_KEY_FOREIGN, ['userid'], 'user', ['id']);
-        $table->add_index('userid_action', XMLDB_INDEX_UNIQUE, ['userid', 'action']);
+        $table->add_index('userid_action', (int) XMLDB_INDEX_UNIQUE, ['userid', 'action']);
 
         if (!$dbman->table_exists($table)) {
             $dbman->create_table($table);
